test: 블록체인 서비스 활용 예제 테스트 작성 - 실제 Build할때는 제외되어야합니다. 단순 참고용

+++ BE/src/main/java/com/ssafy/b105/dto/blockchain/ContractRequestDto.java
@@ -8,7 +8,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class ContractRequestDto {
 
-  private final BigInteger deadLine;
-  private final BigInteger targetAmount;
+  private final Long deadLine;
+  private final Long targetAmount;
 
 }
